+++ packages/workspace/src/browser/workspace-frontend-contribution.spec.ts
@@ -17,6 +17,12 @@
 import { enableJSDOM } from '@theia/core/lib/browser/test/jsdom';
 let disableJSDOM = enableJSDOM();
 
+import { FrontendApplicationConfigProvider } from '@theia/core/lib/browser/frontend-application-config-provider';
+import { ApplicationProps } from '@theia/application-package/lib/application-props';
+FrontendApplicationConfigProvider.set({
+    ...ApplicationProps.DEFAULT.frontend.config
+});
+
 import { expect } from 'chai';
 import { OS } from '@theia/core/lib/common/os';
 import { OpenFileDialogProps } from '@theia/filesystem/lib/browser/file-dialog';